# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AlbumOut": {
        "properties": {
          "albumArtist": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtistRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "coverImageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverimageid"
          },
          "durationMs": {
            "title": "Durationms",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trackCount": {
            "title": "Trackcount",
            "type": "integer"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "title",
          "year",
          "albumArtist",
          "coverImageId",
          "trackCount",
          "durationMs"
        ],
        "title": "AlbumOut",
        "type": "object"
      },
      "AlbumPageOut": {
        "properties": {
          "album": {
            "$ref": "#/components/schemas/AlbumOut"
          },
          "images": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ImageData"
            },
            "title": "Images",
            "type": "object"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackOut"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "album",
          "tracks",
          "images"
        ],
        "title": "AlbumPageOut",
        "type": "object"
      },
...
      "ArtistOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "imageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageid"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sortName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sortname"
          }
        },
        "required": [
          "id",
          "name",
          "sortName",
          "imageId"
        ],
        "title": "ArtistOut",
        "type": "object"
      },
      "ArtistPageOut": {
        "properties": {
          "albums": {
            "items": {
              "$ref": "#/components/schemas/AlbumOut"
            },
            "title": "Albums",
            "type": "array"
          },
          "appearsOn": {
            "items": {
              "$ref": "#/components/schemas/TrackOut"
            },
            "title": "Appearson",
            "type": "array"
          },
          "artist": {
            "$ref": "#/components/schemas/ArtistOut"
          },
          "images": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ImageData"
            },
            "title": "Images",
            "type": "object"
          },
          "topTracks": {
            "items": {
              "$ref": "#/components/schemas/TrackOut"
            },
            "title": "Toptracks",
            "type": "array"
          },
          "trackCount": {
            "title": "Trackcount",
            "type": "integer"
          }
        },
        "required": [
          "artist",
          "albums",
          "topTracks",
          "appearsOn",
          "trackCount",
          "images"
        ],
        "title": "ArtistPageOut",
        "type": "object"
      },
      "ArtistRef": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "ArtistRef",
        "type": "object"
      },
...
      "GenreCount": {
        "properties": {
          "genre": {
            "title": "Genre",
            "type": "string"
          },
          "tracks": {
            "title": "Tracks",
            "type": "integer"
          }
        },
        "required": [
          "genre",
          "tracks"
        ],
        "title": "GenreCount",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HomeOut": {
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/Counts"
          },
          "images": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ImageData"
            },
            "title": "Images",
            "type": "object"
          },
          "playlists": {
            "items": {
              "$ref": "#/components/schemas/PlaylistOut"
            },
            "title": "Playlists",
            "type": "array"
          },
          "recent": {
            "items": {
              "$ref": "#/components/schemas/TrackOut"
            },
            "title": "Recent",
            "type": "array"
          },
          "recentlyAdded": {
            "items": {
              "$ref": "#/components/schemas/TrackOut"
            },
            "title": "Recentlyadded",
            "type": "array"
          }
        },
        "required": [
          "recent",
          "recentlyAdded",
          "playlists",
          "counts",
          "images"
        ],
        "title": "HomeOut",
        "type": "object"
      },
...
      "JobOut": {
        "properties": {
          "job": {
            "title": "Job",
            "type": "string"
          }
        },
        "required": [
          "job"
        ],
        "title": "JobOut",
        "type": "object"
      },
      "LibrarySummaryOut": {
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/Counts"
          },
          "durationMs": {
            "title": "Durationms",
            "type": "integer"
          },
          "firstAddedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstaddedat"
          },
          "genres": {
            "items": {
              "$ref": "#/components/schemas/GenreCount"
            },
            "title": "Genres",
            "type": "array"
          },
          "lastAddedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastaddedat"
          },
          "playedMs": {
            "title": "Playedms",
            "type": "integer"
          },
          "plays": {
            "title": "Plays",
            "type": "integer"
          }
        },
        "required": [
          "counts",
          "durationMs",
          "firstAddedAt",
          "lastAddedAt",
          "genres",
          "plays",
          "playedMs"
        ],
        "title": "LibrarySummaryOut",
        "type": "object"
      },
...
      "PlayerContextOut": {
        "properties": {
          "audio": {
            "$ref": "#/components/schemas/AudioInfo"
          },
          "credits": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Credits",
            "type": "object"
          },
          "extra": {
            "additionalProperties": true,
            "title": "Extra",
            "type": "object"
          },
          "images": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ImageData"
            },
            "title": "Images",
            "type": "object"
          },
          "lyrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LyricsOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "stats": {
            "$ref": "#/components/schemas/TrackStats"
          },
          "track": {
            "$ref": "#/components/schemas/TrackOut"
          }
        },
        "required": [
          "track",
          "lyrics",
          "credits",
          "audio",
          "stats",
          "images"
        ],
        "title": "PlayerContextOut",
        "type": "object"
      },
...
      "ScanIn": {
        "properties": {
          "accountId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accountid"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "ScanIn",
        "type": "object"
      },
...
      "TrackOut": {
        "properties": {
          "addedAt": {
            "format": "date-time",
            "title": "Addedat",
            "type": "string"
          },
          "album": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album"
          },
          "albumId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Albumid"
          },
          "artistDisplay": {
            "title": "Artistdisplay",
            "type": "string"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/TrackArtist"
            },
            "title": "Artists",
            "type": "array"
          },
          "coverImageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverimageid"
          },
          "discNo": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discno"
          },
          "durationMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationms"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "titleDisplay": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Titledisplay"
          },
          "trackNo": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trackno"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "title",
          "titleDisplay",
          "artistDisplay",
          "artists",
          "albumId",
          "album",
          "year",
          "genre",
          "trackNo",
          "discNo",
          "durationMs",
          "coverImageId",
          "addedAt"
        ],
        "title": "TrackOut",
        "type": "object"
      },
...
      "UploadIn": {
        "properties": {
          "filename": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "size": {
            "exclusiveMinimum": 0.0,
            "maximum": 2147483648.0,
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "sha256",
          "size",
          "filename"
        ],
        "title": "UploadIn",
        "type": "object"
      },
      "UploadOut": {
        "properties": {
          "exists": {
            "default": false,
            "title": "Exists",
            "type": "boolean"
          },
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "offset": {
            "default": 0,
            "title": "Offset",
            "type": "integer"
          },
          "state": {
            "default": "receiving",
            "title": "State",
            "type": "string"
          }
        },
        "title": "UploadOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MusiX",
    "version": "2.0.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v2/albums": {
      "get": {
        "operationId": "albums_api_v2_albums_get",
        "parameters": [
          {
            "description": "comma-separated, ≤ 200",
            "in": "query",
            "name": "ids",
            "required": true,
            "schema": {
              "description": "comma-separated, ≤ 200",
              "title": "Ids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlbumOut"
                  },
                  "title": "Response Albums Api V2 Albums Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Albums",
        "tags": [
          "screens"
        ]
      }
    },
    "/api/v2/albums/{album_id}": {
      "get": {
        "operationId": "album_page_api_v2_albums__album_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "album_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Album Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Album Page",
        "tags": [
          "screens"
        ]
      }
    },
...
    "/api/v2/artists": {
      "get": {
        "operationId": "artists_api_v2_artists_get",
        "parameters": [
          {
            "description": "comma-separated, ≤ 200",
            "in": "query",
            "name": "ids",
            "required": true,
            "schema": {
              "description": "comma-separated, ≤ 200",
              "title": "Ids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtistOut"
                  },
                  "title": "Response Artists Api V2 Artists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Artists",
        "tags": [
          "screens"
        ]
      }
    },
    "/api/v2/artists/{artist_id}/page": {
      "get": {
        "operationId": "artist_page_api_v2_artists__artist_id__page_get",
        "parameters": [
          {
            "in": "path",
            "name": "artist_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artist Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Artist Page",
        "tags": [
          "screens"
        ]
      }
    },
...
    "/api/v2/home": {
      "get": {
        "operationId": "home_api_v2_home_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Home",
        "tags": [
          "screens"
        ]
      }
    },
...
    "/api/v2/library/scan": {
      "post": {
        "operationId": "scan_api_v2_library_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Scan",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v2/library/summary": {
      "get": {
        "operationId": "library_summary_api_v2_library_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibrarySummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Library Summary",
        "tags": [
          "screens"
        ]
      }
    },
...
    "/api/v2/player/context/{track_id}": {
      "get": {
        "operationId": "player_context_api_v2_player_context__track_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Track Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerContextOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Player Context",
        "tags": [
          "screens"
        ]
      }
    },
...
    "/api/v2/tracks": {
      "get": {
        "operationId": "get_tracks_api_v2_tracks_get",
        "parameters": [
          {
            "description": "comma-separated, ≤ 200",
            "in": "query",
            "name": "ids",
            "required": true,
            "schema": {
              "description": "comma-separated, ≤ 200",
              "title": "Ids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrackOut"
                  },
                  "title": "Response Get Tracks Api V2 Tracks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Tracks",
        "tags": [
          "library"
        ]
      }
    },
...
    "/api/v2/uploads": {
      "post": {
        "operationId": "start_upload_api_v2_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Upload",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v2/uploads/{upload_id}": {
      "get": {
        "operationId": "get_upload_api_v2_uploads__upload_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Upload",
        "tags": [
          "library"
        ]
      },
      "patch": {
        "operationId": "upload_chunk_api_v2_uploads__upload_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "upload-offset",
            "required": true,
            "schema": {
              "title": "Upload-Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Chunk",
        "tags": [
          "library"
        ]
      }
    }
  }
}